Abstract control plane test
As we will have different control plane implementation, it would be great to have a common base set of tests for them.
